Shift key translation utilities into xcbq
