+++ libqtile/core/types.py
@@ -1,0 +1,7 @@
+from abc import ABCMeta, abstractmethod
+
+
+class Core(metaclass=ABCMeta):
+    pass
+
+
